+++ gsjson.js
@@ -12,16 +12,18 @@ program
     .option('-b, --beautify', 'Beautify final JSON')
     .option('-s, --credentials <s>', 'Service Account credentials JSON data or file path')
     .option('-t, --token <token>', 'Auth token acquired externally')
-    .option('-y, --tokentype <tokentype>', 'Type of the informed token (defaults to Bearer)')
-    .option('-w, --worksheet <n>', 'Worksheet index or title (defaults to first worksheet, can be repeated)', handlePossibleIntList)
+    .option('-y, --tokentype <tokentype>', 'Type of the informed token (defaults to "Bearer")')
+    .option('-w, --worksheet <n>', 'Worksheet index (zero-based) or title (defaults to first worksheet, can be repeated)', handlePossibleList)
     .option('-a, --all-worksheets', 'Return all worksheets (worksheet option is ignored)')
     .option('-c, --hash <column>', 'Column to hash the final object')
     .option('-m, --property-mode <mode>', 'How to handle property names: "camel" (default), "pascal", "nospace" or "none"', /^(camel|pascal|nospace|none)$/i, 'camel')
     .option('-i, --vertical', 'Use the first column as header')
     .option('-l, --list-only', 'Just list the values in arrays')
     .option('-0, --include-header', 'Include header when using "list-only" option')
-    .option('--ignore-col <n>', 'Column name (Excel-like labels) to be ignored (can be repeated, number are also supported)', handlePossibleIntList)
-    .option('--ignore-row <n>', 'Row number to be ignored (can be repeated)', handlePossibleIntList)
+    .option('--header-start <n>', 'Header start line (auto-detected by default)')
+    .option('--header-size <n>', 'Header lines quantity (defaults to 1)')
+    .option('--ignore-col <n>', 'Column name (Excel-like labels) to be ignored (can be repeated, number are also supported)', handlePossibleList)
+    .option('--ignore-row <n>', 'Row number to be ignored (can be repeated)', handlePossibleList)
     .parse(process.argv)
 
 if (program.args.length < 1) {
@@ -48,16 +50,6 @@ helper.spreadsheetToJson(program)
 .catch(function(err) {
     throw err
 })
-
-function handlePossibleIntValue(val) {
-    if (/^\d+$/.test(val))
-        return parseInt(val, 10)
-    return val
-}
-
-function handlePossibleIntList(val, memo) {
-    return handlePossibleList(handlePossibleIntValue(val), memo)
-}
 
 function handlePossibleList(val, memo) {